+++ src/routers/router.tsx
@@ -5,17 +5,17 @@ import Home from "../routes/Home";
 import BoardListPage from "../routes/BoardListPage";
 import BoardWritePage from "../routes/BoardWritePage";
 import BoardDetailPage from "../routes/BoardDetailPage";
-import ChallengePage from "../routes/ChallengePage";
-import ChallengeDetailPage from "../routes/ChallengeDetailPage";
-import ChallengeCAL from "../routes/ChallengeCAL";
-import LoginPage from "../routes/LoginPage";
-import SignUpPage from "../routes/SignUpPage";
-import SignUpProfilePage from "../routes/SignUpProfilePage";
-import AssetConnectPage from "../routes/AssetConnectPage";
-import AssetLoadingPage from "../routes/AssetLoadingPage";
-import ConnectedAssetPage from "../routes/ConnectedAssetPage";
-import TendencyAnalysisPage from "../routes/TendencyAnalysisPage";
-import SignUpSuccessPage from "../routes/SignUpSuccessPage";
+import ChallengePage from "../routes/challenge/ChallengePage";
+import ChallengeDetailPage from "../routes/challenge/ChallengeDetailPage";
+import ChallengeCAL from "../routes/challenge/ChallengeCAL";
+import LoginPage from "../routes/login/LoginPage";
+import SignUpPage from "../routes/signup/SignUpPage";
+import SignUpProfilePage from "../routes/signup/SignUpProfilePage";
+import AssetConnectPage from "../routes/asset/AssetConnectPage";
+import AssetLoadingPage from "../routes/asset/AssetLoadingPage";
+import ConnectedAssetPage from "../routes/asset/ConnectedAssetPage";
+import TendencyAnalysisPage from "../routes/signup/TendencyAnalysisPage";
+import SignUpSuccessPage from "../routes/signup/SignUpSuccessPage";
 // import ProductListPage from "../routes/ProductListPage";
 
 export const mainRouter = [
