+++ src/routes/BoardDetailPage.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import Header from "../components/Header";
-import Navbar from "./Navbar";
+import Navbar from "../components/Navbar";
 import character1 from "../assets/character1.svg";
 import character1_small from "../assets/character1_small.svg";
 import character2 from "../assets/character2.svg";
