+++ src/routes/challenge/ChallengeCAL.tsx
@@ -1,10 +1,10 @@
 import React, { useState } from "react";
 import dayjs from "dayjs";
-import Happy from "../assets/happy-face.svg?react";
-import Sad from "../assets/sad-face.svg?react";
-import Angry from "../assets/angry-face.svg?react";
-import Navbar from "./Navbar";
-import Header from "../components/Header";
+import Happy from "../../assets/happy-face.svg?react";
+import Sad from "../../assets/sad-face.svg?react";
+import Angry from "../../assets/angry-face.svg?react";
+import Navbar from "../../components/Navbar";
+import Header from "../../components/Header";
 
 type DateObject = {
   currentMonth: boolean;
