+++ src/routes/login/LoginPage.tsx
@@ -1,4 +1,4 @@
-import MainLogo from "../assets/main-logo.svg?react";
+import MainLogo from "../../assets/main-logo.svg?react";
 import { Link } from "react-router-dom";
 
 const LoginPage = () => {
